proper error handling

+++ src/user/user.service.ts
@@ -2,13 +2,19 @@ import * as argon from 'argon2';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { UserDto } from './dto';
 import { error } from 'console';
-import { ForbiddenException, Injectable } from '@nestjs/common';
+import { ForbiddenException, HttpException, HttpStatus, Injectable } from '@nestjs/common';
 
 @Injectable()
 export class UserService {
     constructor(private prisma: PrismaService,){}
     async brandName(userId: number, dto:UserDto){
         try {
+            if (!userId)throw new HttpException({
+                status: HttpStatus.BAD_REQUEST,
+                error: 'user id is needed',
+              }, HttpStatus.BAD_REQUEST, {
+                cause: error
+              });
             const user = await this.prisma.user.update({
                 where:{
                     id: userId,
@@ -26,6 +32,12 @@ export class UserService {
     }
     async deleteUser(userId: number,){
         try {
+            if (!userId)throw new HttpException({
+                status: HttpStatus.BAD_REQUEST,
+                error: 'user id is needed',
+              }, HttpStatus.BAD_REQUEST, {
+                cause: error
+              });
             const user = await this.prisma.user.delete({
                 where:{
                     id: userId,
@@ -37,15 +49,30 @@ export class UserService {
         }
     }
     async changePassword(userId: number, newPassword:string, oldpassword:string){
-        if(!newPassword) throw error( "enter your new password")
-        if(!oldpassword) throw  error( "enter your old password")
+        if(!newPassword) throw new HttpException({
+            status: HttpStatus.BAD_REQUEST,
+            error: 'Enter new password',
+          }, HttpStatus.BAD_REQUEST, {
+            cause: error
+          });
+        if(!oldpassword) throw new HttpException({
+            status: HttpStatus.BAD_REQUEST,
+            error: 'Enter old password',
+          }, HttpStatus.BAD_REQUEST, {
+            cause: error
+          });
         try {
             const user = await this.prisma.user.findUnique({
                 where: {
                   id: userId,
                 },
               })
-              if(!user) return ({message: "User not found"})
+              if(!user) throw new HttpException({
+                status: HttpStatus.NOT_FOUND,
+                error: 'user not found',
+              }, HttpStatus.NOT_FOUND, {
+                cause: error
+              });
 
               const pwMatches = await argon.verify(user.hash, oldpassword)
               if(!pwMatches) throw new ForbiddenException('Credentials incorrect')
